+++ SourceryFramework/Sources/Parsing/FileParserType.swift
@@ -9,7 +9,7 @@ public protocol FileParserType {
 
     /// Creates parser for a given contents and path.
     /// - Throws: parsing errors.
-    init(contents: String, forceParse: [String], path: Path?, module: String?) throws
+    init(contents: String, forceParse: [String], parseDocumentation: Bool, path: Path?, module: String?) throws
 
     /// Parses given file context.
     ///
@@ -23,9 +23,9 @@ public enum ParserEngine {
 
 public var parserEngine: ParserEngine = .swiftSyntax
 
-public func makeParser(for contents: String, forceParse: [String] = [], path: Path? = nil, module: String? = nil) throws -> FileParserType {
+public func makeParser(for contents: String, forceParse: [String] = [], parseDocumentation: Bool = false, path: Path? = nil, module: String? = nil) throws -> FileParserType {
     switch parserEngine {
     case .swiftSyntax:
-        return try FileParserSyntax(contents: contents, forceParse: forceParse, path: path, module: module)
+        return try FileParserSyntax(contents: contents, forceParse: forceParse, parseDocumentation: parseDocumentation, path: path, module: module)
     }
 }