+++ SourceryFramework/Sources/Parsing/SwiftSyntax/AST/Class+SwiftSyntax.swift
@@ -20,6 +20,7 @@ extension Class {
           attributes: Attribute.from(node.attributes),
           modifiers: modifiers.map(SourceryModifier.init),
           annotations: annotationsParser.annotations(from: node),
+          documentation: annotationsParser.documentation(from: node),
           isGeneric: node.genericParameterClause?.genericParameterList.isEmpty == false
         )
     }