+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Sourcery CHANGELOG
 
 ## Master
+
+## Fixes
+- correct parsing of rawValue initializer in enum cases, fixes #1010
 ## New Features
 - Added `CLI-Only` subspec to `Sourcery.podspec` [#997](https://github.com/krzysztofzablocki/Sourcery/pull/997)
+- Added documentation comment parsing for all declarations [#1002](https://github.com/krzysztofzablocki/Sourcery/pull/1002)
 - Updates Yams to 4.0.6
+- Enables universal binary
 
 ---